GPT/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2
scikit-learn==1.3.0
numpy==1.26.3
